fix: widen BallPredTracker tick counters to uint64_t (overflow at ~5h)

+++ RLGymPPO_CPP/RLGymSim_CPP/RocketSim/src/Sim/BallPredTracker/BallPredTracker.cpp
@@ -18,11 +18,11 @@ BallPredTracker::~BallPredTracker() {
 void BallPredTracker::UpdatePredFromArena(Arena* arena) {
 	BallState bs = arena->ball->GetState();
 
-	int ticksSinceLastUpdate = arena->tickCount - lastUpdateTickCount;
+	uint64_t ticksSinceLastUpdate = arena->tickCount - lastUpdateTickCount;
 	UpdatePredManual(bs, ticksSinceLastUpdate);
 }
 
-void BallPredTracker::UpdatePredManual(const BallState& curBallState, int ticksSinceLastUpdate) {
+void BallPredTracker::UpdatePredManual(const BallState& curBallState, uint64_t ticksSinceLastUpdate) {
 
 	bool needsFullRepred;
 	if (ticksSinceLastUpdate < predData.size()) {
